- Corrigé le type de retour de ToInt.
git-svn-id: https://svn.opac.ch/svn/cr/branches/vs2005/Epsitec.Cresus@7549 787ff09c-d2b5-7842-b444-8dd0a418631d

+++ Epsitec.Cresus/Common.Types/InvariantConverter.cs
@@ -138,7 +138,7 @@ namespace Epsitec.Common.Types
 				throw new System.ArgumentException ("Value cannot be converted", "obj");
 			}
 		}
-		public static long    ToInt(object obj)
+		public static int     ToInt(object obj)
 		{
 			int value;
